Add FB RSC Entry for React Package (#30310)
I noticed that the www-modern builds pick up the `.experimental.js`
entry points but these flags that are associated with these exports are
not enabled in www, so it gets the wrong builds.

This file is just a clone over the stable `ReactServer.js` one. We
should probably do the reexport pattern instead.